Added content into the add_notebook_tabs area
changed the way its gonna look. now each tweet will be in a short frame,
consisting of the tweet text, the id, and buttons which would allow the
user to say if the tweet is verifiable or not, and if the tweet is
true/false/unverified

Currently not exactly working due to import problems. We'll see how it
goes when we change the way the imports work.