Application successfully connects with twitterWrapper.
Upon notebook tab creation, TwitterWrapper instantiated for a certain
user.
When this happens, it uses a listbox to show the tweets.

TODO: Upon clicking a listing on the listbox, I want to see the text,
its classification and truthfullness on the description frame below.

Known bugs: Application cannot handle emojis above the range "\FFFF"